requirements.txt
python-docx
requests
python-nmap
PyYAML
